04/10/2024

+++ pr-1/TEMPRTUR.C
@@ -1,18 +1,19 @@
+
 #include <stdio.h>
 #include <conio.h>
 
 main ()
 {
-	int c;
+	float c;
 
 	clrscr();
 
 	printf("\n\n");
 
 	printf(" Enter the value of temprature in celsius : ");
-	scanf("%d",&c);
+	scanf("%f",&c);
 
-	printf("\n Temprature in Fahrenheit : %d \n",(c * 9 / 5) + 32 );
+	printf("\n Temprature in Fahrenheit : %.2f \n",(c * 9 / 5) + 32 );
 
 	getch();
